pyproject.toml
[project]
name = "whc"
version = "1.0.0"
description = "Waving Hand Classification."
readme = "README.md"
requires-python = "==3.11.12"
dependencies = [
    "datasets==4.3.0",
    "huggingface-hub==1.0.1",
    "matplotlib==3.10.7",
    "pyarrow==22.0.0",
    "requests==2.32.5",
    "numpy==1.26.4",
    "onnx==1.19.0",
    "onnxruntime-gpu==1.23.0",
    "opencv-python==4.11.0.86",
    "pandas==2.3.3",
    "tqdm==4.67.1",
    "torch==2.5.1",
    "torchvision==0.20.1",
    "tensorboard==2.17.1",
    "onnxsim==0.4.30",
    "sit4onnx==1.0.10",
    "bbalg==1.0.2",
]
